Enable --try-add-blocked-jobs by default
This changes the default behavior to add blocked jobs to a node's batch
if the blocking jobs are already part of that batch.

This is likely the expected behavior from users. It will enable the
workflow where each batch has one aggregation script to upload all data
written to a node's local storage by the main jobs.